reorganize mason
