+++ nvim/home/.config/nvim/lua/zsb/plugins/lsp-config.lua
@@ -1,0 +1,84 @@
+return {
+	{
+		"williamboman/mason.nvim",
+		config = function()
+			require("mason").setup()
+		end,
+	},
+	{
+		"williamboman/mason-lspconfig.nvim",
+		config = function()
+			require("mason-lspconfig").setup({
+				ensure_installed = { "lua_ls" },
+			})
+		end,
+	},
+	{
+		"neovim/nvim-lspconfig",
+		config = function()
+			local capabilities =
+				require("cmp_nvim_lsp").default_capabilities(vim.lsp.protocol.make_client_capabilities())
+
+			local function organize_imports()
+				local params = {
+					command = "_typescript.organizeImports",
+					arguments = { vim.api.nvim_buf_get_name(0) },
+					title = "",
+				}
+				vim.lsp.buf.execute_command(params)
+			end
+
+			local on_attach = function(_, bufnr)
+				local opts = { noremap = true, silent = true }
+
+				vim.api.nvim_buf_set_keymap(bufnr, "n", "gD", "<cmd>lua vim.lsp.buf.declaration()<CR>", opts)
+				vim.api.nvim_buf_set_keymap(bufnr, "n", "gd", "<cmd>lua vim.lsp.buf.definition()<CR>", opts)
+				vim.api.nvim_buf_set_keymap(bufnr, "n", "K", "<cmd>lua vim.lsp.buf.hover()<CR>", opts)
+				vim.api.nvim_buf_set_keymap(bufnr, "n", "gi", "<cmd>lua vim.lsp.buf.implementation()<CR>", opts)
+				vim.api.nvim_buf_set_keymap(bufnr, "n", "<C-k>", "<cmd>lua vim.lsp.buf.signature_help()<CR>", opts)
+				vim.api.nvim_buf_set_keymap(bufnr, "n", "gr", "<cmd>lua vim.lsp.buf.references()<CR>", opts)
+				vim.api.nvim_buf_set_keymap(bufnr, "n", "<leader>q", "<cmd>lua vim.diagnostic.setloclist()<CR>", opts)
+			end
+
+			local defaultSetUp = {
+				on_attach = on_attach,
+				capabilities = capabilities,
+			}
+
+			local lspconfig = require("lspconfig")
+
+			lspconfig["lua_ls"].setup(Extend(defaultSetUp, {
+				settings = {
+					Lua = {
+						diagnostics = {
+							globals = { "vim", "hs", "table", "kset", "createCmd" },
+						},
+					},
+				},
+			}))
+
+			lspconfig["tsserver"].setup(Extend(defaultSetUp, {
+				commands = {
+					OrganizeImports = {
+						organize_imports,
+						description = "Organize Imports",
+					},
+				},
+			}))
+
+			local defaultConfigServers = {
+				"pyright",
+				"cssls",
+				"emmet_ls",
+				"tailwindcss",
+				"bashls",
+				"gopls",
+				"rust_analyzer",
+			}
+
+			for _, server in pairs(defaultConfigServers) do
+				lspconfig[server].setup(defaultSetUp)
+			end
+		end,
+	},
+}
